--- a/app/client/src/components/pages/Community.Tabs.Overview.js
+++ b/app/client/src/components/pages/Community.Tabs.Overview.js
@@ -634,7 +634,7 @@ function MonitoringAndSensorsTab({
 
     setNormalizedMonitoringLocations(stations);
 
-    plotStations(stations, monitoringLocationsLayer, services);
+    plotStations(stations, monitoringLocationsLayer);
   }, [monitoringLocations.data, monitoringLocationsLayer, services]);
 
   const allMonitoringAndSensors = [