… app/client/src/components/shared/LocationMap.js
… app/client/src/components/shared/VirtualizedList.js
… app/client/src/contexts/locationSearch.js
… app/client/src/utils/hooks.js
… app/client/src/utils/mapFunctions.js